Fix sign message to resolve issue
This PR fixes the issue #1. The issue was caused by the fact that Massa Station has 2 endpoints sign (to sign operation) and signMessage (to sign a message).
To fix the current implementation we provide a dumb description and set DisplayData to true but this
feature must be implemented in the future.